pyproject.toml
[project]
name = "xhcart-core"
version = "0.1.0"
description = "XHGC Cart Packaging Tool"
authors = [
    { name = "Nixie Studio", email = "contact@nixie.studio" }
]
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.9"

[project.optional-dependencies]
image = [
    "Pillow>=10"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]